# allocate small strings in 8-byte-aligned pools