This is a less-disruptive alternative to #40727 that keeps the representation the same and just uses less alignment padding.

Before:
```
function example()
    for j = 1:20, i = 1:1_000_000
        string(i)
    end
end

julia> @btime example()
  598.121 ms (40000000 allocations: 1.79 GiB)

julia> @btime for i = 1:2^24
    "a"*"b"
end
  247.768 ms (16777216 allocations: 512.00 MiB)
```

after:
```
julia> @btime example()
  607.298 ms (40000000 allocations: 1.64 GiB)

julia> @btime for i = 1:2^24
    "a"*"b"
end
  247.473 ms (16777216 allocations: 384.00 MiB)
```

@nanosoldier `runbenchmarks("string", vs=":master")`