@nanosoldier 2021-06-16
[Your benchmark job](https://github.com/JuliaLang/julia/pull/41247) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/4e1ca7a_vs_dd94ceb/report.md). cc @christopher-dG