@JeffBezanson 2021-06-18
We discussed this a bit on triage. Seems like we agree it's a win-win. Plus we could use the new pools in the future for other objects with less-strict alignment, e.g. a struct of two bytes.