@oscardssmith 2021-06-25
If we're back-porting the change to array size to 1.7, do we want to back-port this too?